Her har vi nå satt slik at formen legges til i databasen, slik at vi kan se den i rapports tabellen i admin siden - Emil

+++ Databases/Tables/Marker.cs
@@ -5,8 +5,8 @@
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Lattitude { get; set; }
-        public int Longitude { get; set; }
+        public double Lattitude { get; set; }
+        public double Longitude { get; set; }
         public int Altitude { get; set; }
     }
 }
